requirements.txt
requests>=2.32.0
feedparser>=6.0.11
python-dotenv>=1.0.1
openai>=1.40.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
